python_ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
firebase-admin==6.2.0
google-cloud-aiplatform==1.42.0
google-generativeai==0.3.0
aiohttp==3.9.1
python-multipart==0.0.6
